// src/app/models/ProgramaValores.ts
export class ProgramaValores {
  n_id_programa_mensual: number;
  n_id_ofic: number;
  v_n_id_contrati: number;
  v_idacti: string;
  v_idsubacti_1:string;
	v_idsubacti_2:string;
  n_val_prog_mensual: number;
  n_id_estado: number;
  d_periodo: Date;
  v_periodo: string;
  n_val_prog_total: number;
  n_cant_periodos: number;
  a_v_usucre: string;
  a_d_feccre: Date;
  a_v_usumod: string;
  a_d_fecmod: Date;
  v_v_mes_inicio: number;
  v_v_anio_inicio: number;
  v_editable: number;
  }